Migrate extension from Ruff to ty (#4)
## Summary

This PR adds a VS Code extension for `ty`, modeled here as a diff
against the Ruff extension.

The major changes are as follows:

1. I removed the tests for now, since they're specific to the linter and
formatter. We can re-add tests later.
2. I removed all irrelevant settings. For now, we only support a small
number of settings, mostly related to discovering Python (as opposed to
configuring `ty`'s own behavior).
3. I removed all code related to the legacy LSP (a lot!).
4. I removed the `CONTRIBUTING.md`.

I retained the Python dependency for now, since we need it to discover
the `ty` executable in the selected virtual environment.