Avoid starting the language server if interpreter unavailable (#29)
## Summary

This PR fixes a bug which avoid return an error when a Python
interpreter was not selected (as reported on Discord). This would happen
if `ty.interpreter` is empty and there's no local or global Python
interpreter available. This shouldn't be required once #5 is
implemented.

## Test Plan

After this PR, the VS Code extension will provide the "error" log and
ask users to select a Python interpreter and not start the language
server.

<img width="1724" alt="Screenshot 2025-05-17 at 11 26 21"
src="https://github.com/user-attachments/assets/9e66a444-ac50-48ca-8223-3da4f75c8d5b"
/>