Add `python.ty.disableLanguageServices` config (#36)
## Summary

PR adding support for it in the language server:
https://github.com/astral-sh/ruff/pull/18230

This PR adds support for `python.ty.disableLanguageServices` in the
extension to extract the value and pass it as server setting.

Closes: #20 

## Test Plan

The video starts by the commented out
`python.ty.disableLanguageServices` and you can see that there are inlay
hints, hover support, diagnostics, etc.

The hover content for `x` is `Literal[1]`.

Once it's set to true, the server restarts and inlay hints are gone,
hover content is gone but the diagnostics still remain as expected.


https://github.com/user-attachments/assets/9280e7b0-bf89-4d99-bfa9-d10a6d3368f7