Fix release workflow (#122)
## Summary

This PR fixes the release workflow again.

It makes the following changes:
- Take as input for workflow dispatch as to whether this is a stable or
prerelease
- Use the `${{ github.event.release.prerelease }}` check under `if`
instead of the bash script. This is important because the steps are run
in various operating systems where one is Windows which cannot run bash

Regarding the second bullet point, I think this must be the reason that
the steps were divided between stable and prerelease workflow and I've
reverted back to that structure.

## Test Plan

Going to merge this and manually release.