Add more links to the README (#210)
## Summary

I decided recently it was finally time to dogfood ty to myself in
VSCode. Two things confused me while installing the extension:
- I searched for "ty-vscode" in the marketplace, and no results came up.
The reason is that the extension is actually called "ty", not
"ty-vscode". I've added a link to the VSCode marketplace so that this is
obvious to anybody reading the README.
- The README recommends to edit your `settings.json` file, but I'm a
VSCode noob and had no idea how to do that. I've added a link to the
`settings.json` reference so that other people in a similar boat don't
have to look it up manually