Exclude extension specific settings from workspace configuration response (#237)
This PR updates the middleware to remove all the extension specific
settings from the response of `workspace/configuration` request.
Otherwise, the server would consider them as unknown options and raise a
warning.

This PR was written with the help of Claude but I tested it in VS Code
myself by specifying all of these options before and after to make sure
a warning isn't raised.